refactor: applied PR feedback
