+++ src/utils/whitelist-cookies.js
@@ -8,6 +8,9 @@ Cypress.Cookies.defaults({
     // We assume that the cookie is never splitted into more than 10 parts
     ...Array(10)
       .fill(1)
-      .map((x, y) => `${Cypress.env('auth0SessionCookieName')}.${x + y}`),
+      .map(
+        (value, index) =>
+          `${Cypress.env('auth0SessionCookieName')}.${value + index}`,
+      ),
   ],
 });
